+++ REQUIRE
@@ -9,3 +9,4 @@ DataFrames
 Docile
 Grid
 GaussianMixtures
+SloanDigitalSkySurvey
